type-c-service/driver/tps6699x: Implement status caching
In certain cases, reading the port status may reveal additional events
that need to be signaled to the type-C service. To avoid notifying
the type-C service excessively this commit adds caching of the PD status
which introduces a layer of abstraction where additional events can be
checked before notifying the type-C service.